refactor: refactor query schema
