+++ src/routes/graphql/root-schema.ts
@@ -1,0 +1,10 @@
+import { GraphQLSchema } from 'graphql/index.js';
+import { createQuerySchema } from './create-query-schema.js';
+
+
+
+export const createScheme = ()  => {
+  return new GraphQLSchema({
+    query: createQuerySchema(),
+  });
+}
